* add some debugging output to prelink failures * change all size_t (for coords) -> int32_t, to circumvent a bunch of signedness mismatch errors

+++ aspng_surface.cpp
@@ -9,15 +9,15 @@ std::ostream &operator<<(std::ostream &os, const Rgb &rgb) {
     return os;
 }
 
-bool AspngSurface::contains(AspngSurface *other, size_t x, size_t y) {
+bool AspngSurface::contains(AspngSurface *other, int32_t x, int32_t y) {
     if (x + other->get_width() >= this->get_width()) {
         return false;
     }
     if (y + other->get_height() >= this->get_height()) {
         return false;
     }
-    for (size_t u = 0; u < other->get_width(); u++) {
-        for (size_t v = 0; v < other->get_height(); v++) {
+    for (int32_t u = 0; u < other->get_width(); u++) {
+        for (int32_t v = 0; v < other->get_height(); v++) {
             if (this->get_pixel(x + u, y + v) != other->get_pixel(u, v)) {
                 return false;
             }
@@ -26,11 +26,11 @@ bool AspngSurface::contains(AspngSurface *other, size_t x, size_t y) {
     return true;
 }
 
-void AspngSurface::copy(AspngSurface *other, size_t x, size_t y) {
+void AspngSurface::copy(AspngSurface *other, int32_t x, int32_t y) {
     ASSERT(x + other->get_width() < this->get_width());
     ASSERT(y + other->get_height() < this->get_height());
-    for (size_t u = 0; u < other->get_width(); u++) {
-        for (size_t v = 0; v < other->get_height(); v++) {
+    for (int32_t u = 0; u < other->get_width(); u++) {
+        for (int32_t v = 0; v < other->get_height(); v++) {
             this->set_pixel(x + u, y + v, other->get_pixel(u, v));
         }
     }
