+++ aspng_surface.h
@@ -19,14 +19,14 @@ std::ostream &operator<<(std::ostream &, const Rgb &);
 
 class AspngSurface {
 public:
-    virtual Rgb get_pixel(size_t, size_t) = 0;
-    virtual void set_pixel(size_t, size_t, Rgb) = 0;
-    virtual size_t get_width(void) = 0;
-    virtual size_t get_height(void) = 0;
+    virtual Rgb get_pixel(int32_t, int32_t) = 0;
+    virtual void set_pixel(int32_t, int32_t, Rgb) = 0;
+    virtual int32_t get_width(void) = 0;
+    virtual int32_t get_height(void) = 0;
     virtual void start_draw(void);
     virtual void finish_draw(void);
-    bool contains(AspngSurface *, size_t, size_t);
-    void copy(AspngSurface *, size_t, size_t);
+    bool contains(AspngSurface *, int32_t, int32_t);
+    void copy(AspngSurface *, int32_t, int32_t);
 };
 
 #endif
