+++ led_device.cpp
@@ -17,9 +17,9 @@ std::string LEDDevice::template_name(void) {
 }
 
 // A valid LEDDevice must be empty (black).
-bool LEDDevice::sub_parse(AspngSurface *surface, size_t min_x, size_t min_y, size_t max_x, size_t max_y) {
-    for (size_t x = min_x; x <= max_x; x++) {
-        for (size_t y = min_y; y <= max_y; y++) {
+bool LEDDevice::sub_parse(AspngSurface *surface, int32_t min_x, int32_t min_y, int32_t max_x, int32_t max_y) {
+    for (int32_t x = min_x; x <= max_x; x++) {
+        for (int32_t y = min_y; y <= max_y; y++) {
             if (surface->get_pixel(x, y) != Rgb(0, 0, 0)) {
                 return false;
             }
@@ -63,10 +63,10 @@ std::list<Patch *> LEDDevice::sub_patches(void) {
 }
 
 // A nice neon green color if active, otherwise dark.
-void LEDDevice::sub_draw(AspngSurface *surface, size_t min_x, size_t min_y, size_t max_x, size_t max_y) {
+void LEDDevice::sub_draw(AspngSurface *surface, int32_t min_x, int32_t min_y, int32_t max_x, int32_t max_y) {
     Rgb color = this->active ? Rgb(0, 0xff, 0) : Rgb(0, 0x30, 0);
-    for (size_t x = min_x; x <= max_x; x++) {
-        for (size_t y = min_y; y <= max_y; y++) {
+    for (int32_t x = min_x; x <= max_x; x++) {
+        for (int32_t y = min_y; y <= max_y; y++) {
             surface->set_pixel(x, y, color);
         }
     }
