+++ simple_aspng_surface.cpp
@@ -3,16 +3,16 @@
 #include "common.h"
 #include "simple_aspng_surface.h"
 
-SimpleAspngSurface::SimpleAspngSurface(size_t w, size_t h) {
+SimpleAspngSurface::SimpleAspngSurface(int32_t w, int32_t h) {
     this->width = w;
     this->height = h;
     this->pixels.resize(w * h);
 }
 
-void SimpleAspngSurface::resize(size_t w, size_t h) {
+void SimpleAspngSurface::resize(int32_t w, int32_t h) {
     std::vector<Rgb> new_pixels(w * h);
-    for (size_t x = 0; x < this->width; x++) {
-        for (size_t y = 0; y < this->height; y++) {
+    for (int32_t x = 0; x < this->width; x++) {
+        for (int32_t y = 0; y < this->height; y++) {
             if ((x < w) && (y < h)) {
                 new_pixels[(w * y) + x] = this->pixels[(this->width * y) + x];
             }
@@ -23,21 +23,21 @@ void SimpleAspngSurface::resize(size_t w, size_t h) {
     this->height = h;
 }
 
-size_t SimpleAspngSurface::get_width(void) {
+int32_t SimpleAspngSurface::get_width(void) {
     return this->width;
 }
 
-size_t SimpleAspngSurface::get_height(void) {
+int32_t SimpleAspngSurface::get_height(void) {
     return this->height;
 }
 
-Rgb SimpleAspngSurface::get_pixel(size_t x, size_t y) {
+Rgb SimpleAspngSurface::get_pixel(int32_t x, int32_t y) {
     ASSERT((x >= 0) && (x < this->width));
     ASSERT((y >= 0) && (y < this->height));
     return this->pixels[(this->width * y) + x];
 }
 
-void SimpleAspngSurface::set_pixel(size_t x, size_t y, Rgb rgb) {
+void SimpleAspngSurface::set_pixel(int32_t x, int32_t y, Rgb rgb) {
     ASSERT((x >= 0) && (x < this->width));
     ASSERT((y >= 0) && (y < this->height));
     this->pixels[(this->width * y) + x] = rgb;
